+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGELOG.md
 
 ## 2.1.1 - 2022-05-13
-### Features (1 change)
+### Features (2 changes)
 - Top elements in product and content search result
+- Custom search query in categories
 
 ### Fix (1 change)